requirements.txt
# Google Agent Development Kit (ADK)
google-adk>=1.0.0

# LiteLLM for Mistral AI integration
litellm>=1.0.0

# python-dotenv for environment variable management
python-dotenv>=1.0.0
